Oh god that's sexy. STL does all the heavy lifting, leaving me with nothing but elegance!

+++ Test/Utilities.h
@@ -39,5 +39,5 @@ namespace Test
 		});
 	}
 
-	int GetNumberOfFailedTests(const OuterTestList& tests);
+	size_t GetNumberOfFailedTests(const OuterTestList& tests);
 }
